requirements.txt
sentence_transformers
requests
bs4
tqdm
faiss-cpu